+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = "ihnatiev-site-v0.6.19c4-r1";
+const CACHE_NAME = "ihnatiev-site-v0.6.19c5-r1";
 const APP_SHELL = [
   "./",
   "./index.html",
@@ -13,6 +13,7 @@ const APP_SHELL = [
   "./js/gallery-lightbox.js",
   "./js/document-lightbox.js",
   "./js/downloads-renderer.js",
+  "./js/youtube-feed.js",
   "./app.js",
   "./manifest.webmanifest",
   "./files/media/icon-192.png",